Change create to use body (#4)
* Changed the api to accept a json body instead of in the route itself

* Adding some todos

* Changed the api

* Changed the params.id to id

* Turned off dynamic nav bar to see if it will fix itself

* Changed the nav bar

* Made sure body is 100%

* Started on auth page

+++ app/api/pools/route.js
@@ -2,7 +2,29 @@ import { NextResponse } from "next/server"
 import prisma from "../../libs/prismadb";
 
 export async function GET(request) {
-  const pools = await prisma.pool.findMany({
+  const id = request.nextUrl.searchParams.get("id");
+
+  if (!id) {
+    const pools = await prisma.pool.findMany({
+      include: {
+        teams: true,
+        matches: {
+          include: {
+            team1: true,
+            team2: true,
+            workTeam: true,
+            sets: true
+          }
+        },
+      }
+    })
+  
+    return NextResponse.json(pools)
+  }
+  const pool = await prisma.pool.findUnique({
+    where: {
+      id: id
+    },
     include: {
       teams: true,
       matches: {
@@ -16,14 +38,44 @@ export async function GET(request) {
     }
   })
 
-  return NextResponse.json(pools)
+  if (!pool) {
+    return new NextResponse("Need Pool Name", { status: 400 })
+  }
+  return NextResponse.json(pool)
+}
+
+export async function POST(request, { params }) {
+  const { name } = await request.json()
+
+  if (!name) {
+    return new NextResponse("Need Pool Name", { status: 400 })
+  }
+
+  const pool = await prisma.pool.create({
+    data: {
+      name: name,
+    }
+  }) 
+
+  return NextResponse.json(pool)
 }
 
 export async function DELETE(request) {
+  const id = request.nextUrl.searchParams.get("id");
 
-  const pools = await prisma.pool.deleteMany({})
+  if (!id) {
+    const pools = await prisma.pool.deleteMany({})
 
-  return NextResponse.json(pools)
+    return NextResponse.json(pools)
+  }
+
+  const pool = await prisma.pool.delete({
+    where: {
+      id: id
+    }
+  })
+
+  return NextResponse.json(pool)
 
 }
